The codecomment is now locked to scroll window too
Now it won't go off the document page or the scroll viewport.  Yay!

Used jquery's position() from http://api.jquery.com/position/, which
``The .position() method allows us to retrieve the current position of
an element relative to the offset parent. Contrast this with .offset(),
which retrieves the current position relative to the document. When
positioning a new element near another one and within the same
containing DOM element, .position() is the more useful.''